requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.3
python-multipart>=0.0.9
openai>=1.12.0
google-genai>=1.0.0
yfinance>=0.2.36
requests>=2.31.0
aiohttp>=3.9.3
feedparser>=6.0.11
beautifulsoup4>=4.12.3
newspaper3k>=0.2.8
nltk>=3.8.1
textblob>=0.18.0
vaderSentiment>=3.3.2
python-dotenv>=1.0.0
pydantic>=2.6.1
pydantic-settings>=2.1.0
click>=8.1.7
rich>=13.7.0
schedule>=1.2.1
tenacity>=8.2.3
httpx>=0.26.0
plotly>=5.18.0
pandas>=2.2.0
numpy>=1.26.4
